requirements.txt
aiogram==2.19
aiohttp==3.8.1; python_version >= "3.6"
aioredis==2.0.1; python_version >= "3.6"
aiosignal==1.2.0; python_version >= "3.6"
async-timeout==4.0.2; python_version >= "3.6"
attrs==21.4.0; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.6"
babel==2.9.1; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.4.0"
cachetools==5.0.0; python_version >= "3.7" and python_version < "4.0" and python_full_version >= "3.6.0" and (python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6")
certifi==2021.10.8; python_full_version >= "3.6.0" and python_version >= "3.6"
charset-normalizer==2.0.12; python_full_version >= "3.6.0" and python_version >= "3.6"
environs==9.5.0; python_version >= "3.6"
frozenlist==1.3.0; python_version >= "3.7"
google-api-core==2.7.1; python_version >= "3.6"
google-api-python-client==2.40.0; python_version >= "3.6"
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.0; python_version >= "3.6"
google-auth==2.6.0; python_full_version >= "3.6.0" and python_version >= "3.6" and (python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6")
googleapis-common-protos==1.55.0; python_version >= "3.6"
greenlet==1.1.2; python_version >= "3" and python_full_version < "3.0.0" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32") and (python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0") or python_version >= "3" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32") and (python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0") and python_full_version >= "3.5.0"
gspread-asyncio==1.5.3; python_version >= "3.6"
gspread-formatting==1.0.6
gspread==4.0.1; python_full_version >= "3.4.0"
httplib2==0.20.4; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.6"
idna==3.3; python_full_version >= "3.6.0" and python_version >= "3.6"
marshmallow==3.14.1; python_version >= "3.6"
multidict==6.0.2; python_version >= "3.7"
oauth2client==4.1.3
oauthlib==3.2.0; python_version >= "3.6" and python_full_version >= "3.4.0"
protobuf==3.19.4; python_version >= "3.6"
pyasn1-modules==0.2.8; python_full_version >= "3.6.0" and python_version >= "3.6" and (python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6")
pyasn1==0.4.8; python_full_version >= "3.6.0" and python_version >= "3.6" and (python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6") and python_version < "4"
pyparsing==3.0.7; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.6"
python-dotenv==0.19.2; python_version >= "3.6"
pytz==2021.3; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.4.0"
requests-oauthlib==1.3.1; python_version >= "3.6" and python_full_version >= "3.4.0"
requests==2.27.1; python_full_version >= "3.6.0" and python_version >= "3.6"
rsa==4.8; python_version >= "3.6" and python_version < "4" and python_full_version >= "3.6.0" and (python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6")
six==1.16.0; python_full_version >= "3.6.0" and python_version >= "3.6" and (python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6") and (python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.3.0" and python_version >= "3.6")
sqlalchemy==1.4.32; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.6.0")
typing-extensions==4.1.1; python_version >= "3.6"
uritemplate==4.1.1; python_version >= "3.6"
urllib3==1.26.8; python_full_version >= "3.6.0" and python_version < "4" and python_version >= "3.6"
yarl==1.7.2; python_version >= "3.6"
